+++ tests/unit/specs/components/config/channels/telegram/Config.spec.js
@@ -48,9 +48,6 @@ describe('TelegramConfig.vue', () => {
       propsData: {
         app: singleApp,
       },
-      mocks: {
-        $t: () => 'some specific text',
-      },
     });
   });
