+++ tests/unit/specs/components/config/channels/whatsapp/components/tabs/ProfileTab.spec.js
@@ -25,6 +25,7 @@ import VueRouter from 'vue-router';
 import { shallowMount, createLocalVue } from '@vue/test-utils';
 import ProfileTab from '@/components/config/channels/whatsapp/components/tabs/ProfileTab.vue';
 import { singleApp } from '../../../../../../../../__mocks__/appMock';
+import storeMock from '../../../../../../store/appType/whatsapp/__mocks__/store.js';
 import i18n from '@/utils/plugins/i18n';
 
 const localVue = createLocalVue();
@@ -33,18 +34,28 @@ localVue.use(Vuex);
 describe('whatsapp/components/tabs/ProfileTab.vue', () => {
   let wrapper;
   let actions;
+  let getters;
   let store;
 
   beforeEach(() => {
     actions = {
       updateWppProfile: jest.fn(),
       deleteWppProfilePhoto: jest.fn(),
+      fetchWppContactInfo: jest.fn(),
+      updateWppContactInfo: jest.fn(),
+    };
+
+    getters = {
+      loadingContactInfo: jest.fn(),
+      fetchedContactInfo: jest.fn(),
+      contactInfo: jest.fn(() => storeMock.mockedContactInfo),
     };
 
     store = new Vuex.Store({
       modules: {
         WhatsApp: {
           namespaced: true,
+          getters,
           actions,
         },
       },
@@ -54,9 +65,6 @@ describe('whatsapp/components/tabs/ProfileTab.vue', () => {
       localVue,
       store,
       i18n,
-      mocks: {
-        $t: () => 'some specific text',
-      },
       stubs: {
         DynamicForm: true,
         UnnnicButton: true,
@@ -84,22 +92,67 @@ describe('whatsapp/components/tabs/ProfileTab.vue', () => {
     expect(wrapper).toMatchSnapshot();
   });
 
-  describe('updateInputs()', () => {
+  describe('mounted()', () => {
+    it('should call unnnicCallAlert if fetchWppContactInfo fails', () => {
+      actions.fetchWppContactInfo.mockImplementation(() => {
+        throw new Error('error');
+      });
+      wrapper = shallowMount(ProfileTab, {
+        localVue,
+        i18n,
+        store,
+        propsData: {
+          app: singleApp,
+          profile: {
+            photoFile: 'photo',
+            status: 'status',
+            business: {
+              description: 'description',
+              vertical: 'option1',
+              vertical_choices: ['option1', 'option2'],
+            },
+          },
+        },
+        mocks: {
+          $t: () => 'some specific text',
+        },
+        stubs: {
+          DynamicForm: true,
+          UnnnicButton: true,
+        },
+      });
+      expect(mockUnnnicCallAlert).toHaveBeenCalled();
+    });
+  });
+
+  describe('updateProfileInputs()', () => {
     it('should set profile info input data correctly', async () => {
       const inputData = {
         index: 0,
         value: 'value',
       };
       expect(wrapper.vm.profileInputs[inputData.index].value).not.toEqual(inputData.value);
-      wrapper.vm.updateInputs(inputData);
+      wrapper.vm.updateProfileInputs(inputData);
       expect(wrapper.vm.profileInputs[inputData.index].value).toEqual(inputData.value);
+    });
+  });
+
+  describe('updateContactInfoInputs()', () => {
+    it('should set contact info input data correctly', async () => {
+      const inputData = {
+        index: 0,
+        value: 'value',
+      };
+      expect(wrapper.vm.contactInfoInputs[inputData.index].value).not.toEqual(inputData.value);
+      wrapper.vm.updateContactInfoInputs(inputData);
+      expect(wrapper.vm.contactInfoInputs[inputData.index].value).toEqual(inputData.value);
     });
   });
 
   describe('profileInputs', () => {
     it('should set modifiedInitialPhoto as true if profile_image was modified', async () => {
       expect(wrapper.vm.modifiedInitialPhoto).toBeFalsy();
-      wrapper.vm.updateInputs({ index: 0, value: [{ lastModified: 123 }] });
+      wrapper.vm.updateProfileInputs({ index: 0, value: [{ lastModified: 123 }] });
       await wrapper.vm.$nextTick();
       expect(wrapper.vm.modifiedInitialPhoto).toBeTruthy();
     });
@@ -182,9 +235,6 @@ describe('whatsapp/components/tabs/ProfileTab.vue', () => {
         localVue,
         store,
         i18n,
-        mocks: {
-          $t: () => 'some specific text',
-        },
         stubs: {
           DynamicForm: true,
           UnnnicButton: true,
@@ -220,15 +270,71 @@ describe('whatsapp/components/tabs/ProfileTab.vue', () => {
       );
     });
 
+    describe('getInputValue()', () => {
+      it('should return value from contactInfoInputs based on input name', () => {
+        const inputName = 'address';
+        expect(wrapper.vm.getContactInfoInputValue(inputName)).toEqual(
+          storeMock.mockedContactInfo.address,
+        );
+      });
+    });
+
+    describe('saveContactInfo()', () => {
+      it('should call updateWppContactInfo action', async () => {
+        expect(actions.updateWppContactInfo).not.toHaveBeenCalled();
+        await wrapper.vm.saveContactInfo();
+        expect(actions.updateWppContactInfo).toHaveBeenCalledTimes(1);
+      });
+
+      it('should call unnnicCallAlert on success', async () => {
+        expect(mockUnnnicCallAlert).not.toHaveBeenCalled();
+        await wrapper.vm.saveContactInfo();
+        expect(mockUnnnicCallAlert).toHaveBeenCalledTimes(1);
+        expect(mockUnnnicCallAlert).toHaveBeenCalledWith(
+          expect.objectContaining({
+            props: {
+              title: 'Success',
+              text: expect.any(String),
+              icon: expect.any(String),
+              scheme: expect.any(String),
+              position: expect.any(String),
+              closeText: expect.any(String),
+            },
+            seconds: expect.any(Number),
+          }),
+        );
+      });
+
+      it('should call unnnicCallAlert on error', async () => {
+        actions.updateWppContactInfo.mockImplementation(() => Promise.reject(new Error('error')));
+        expect(mockUnnnicCallAlert).not.toHaveBeenCalled();
+        await wrapper.vm.saveContactInfo();
+        expect(mockUnnnicCallAlert).toHaveBeenCalledTimes(1);
+        expect(mockUnnnicCallAlert).toHaveBeenCalledWith(
+          expect.objectContaining({
+            props: {
+              title: 'Error',
+              text: expect.any(String),
+              icon: expect.any(String),
+              scheme: expect.any(String),
+              position: expect.any(String),
+              closeText: expect.any(String),
+            },
+            seconds: expect.any(Number),
+          }),
+        );
+      });
+    });
+
     it('should call toBase64() if photo is provided', async () => {
-      wrapper.vm.updateInputs({ index: 0, value: ['photo'] });
+      wrapper.vm.updateProfileInputs({ index: 0, value: ['photo'] });
       expect(mockToBase64).not.toHaveBeenCalled();
       await wrapper.vm.saveProfile();
       expect(mockToBase64).toHaveBeenCalledTimes(1);
     });
 
     it('should call isValidPhotoSize() if photo is provided', async () => {
-      wrapper.vm.updateInputs({ index: 0, value: ['photo'] });
+      wrapper.vm.updateProfileInputs({ index: 0, value: ['photo'] });
       mockToBase64.mockImplementation(() => Promise.resolve('b64photo'));
       const spy = spyOn(wrapper.vm, 'isValidPhotoSize');
       expect(spy).not.toHaveBeenCalled();
@@ -237,7 +343,7 @@ describe('whatsapp/components/tabs/ProfileTab.vue', () => {
     });
 
     it('should return early if photo has an invalid size', async () => {
-      wrapper.vm.updateInputs({ index: 0, value: ['photo'] });
+      wrapper.vm.updateProfileInputs({ index: 0, value: ['photo'] });
       await wrapper.setData({ modifiedInitialPhoto: true });
       mockGetHeightAndWidthFromDataUrl.mockImplementation(() => {
         return Promise.resolve({
@@ -250,7 +356,7 @@ describe('whatsapp/components/tabs/ProfileTab.vue', () => {
     });
 
     it('should call updateWppProfile if photo has a valid size', async () => {
-      wrapper.vm.updateInputs({ index: 0, value: ['photo'] });
+      wrapper.vm.updateProfileInputs({ index: 0, value: ['photo'] });
       await wrapper.setData({ modifiedInitialPhoto: true });
       mockGetHeightAndWidthFromDataUrl.mockImplementation(() => {
         return Promise.resolve({
@@ -263,7 +369,7 @@ describe('whatsapp/components/tabs/ProfileTab.vue', () => {
     });
 
     it('should call unnnicCallAlert in error state if request fails', async () => {
-      wrapper.vm.updateInputs({ index: 0, value: ['photo'] });
+      wrapper.vm.updateProfileInputs({ index: 0, value: ['photo'] });
       await wrapper.setData({ modifiedInitialPhoto: true });
       mockGetHeightAndWidthFromDataUrl.mockImplementation(() => {
         return Promise.resolve({
